pyproject.toml
[project]
name = "plaud-processor"
version = "1.0.0"
description = "AI-Powered Meeting Processor - Transform audio recordings into structured Obsidian notes"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "Peter Bobov" }
]
keywords = ["audio", "transcription", "whisper", "obsidian", "meeting-notes", "openai"]

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "librosa>=0.10.0",
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
    "pyannote-audio>=3.1.0",
    "python-dateutil>=2.8.0",
    "typing-extensions>=4.0.0",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]
